week10_production_observability/requirements.txt
langgraph>=0.2.60
langgraph-checkpoint-sqlite>=2.0.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
langsmith>=0.1.130
faker>=30.0.0
python-dotenv>=1.0.0
anthropic>=0.40.0
claude-agent-sdk>=0.1.0
ipykernel>=6.29.0
jupyter>=1.1.0
jupytext>=1.16.0
grandalf>=0.8          # needed for app.get_graph().draw_ascii()
